tools: add commandline flags for explicit version info parser selection
instead of trying to detect the appropriate parser, the desired
implementation can be selected explicitly via commandline arguments.
this is mainly intended for scenarios where multiple parsers could
be selected, but only the user known which one is more accurate.
alternatively, the version can be provided via commandline arguments
as well using the `mock` parser implementation.